+++ src/backend/marsha/core/api/__init__.py
@@ -3,6 +3,11 @@
 from .account import *  # noqa isort:skip
 from .base import *  # noqa isort:skip
 from .file import *  # noqa isort:skip
+from .live_registration import *  # noqa isort:skip
+from .pairing_challenge import *  # noqa isort:skip
 from .playlist import *  # noqa isort:skip
+from .shared_live_media import *  # noqa isort:skip
+from .thumbnail import *  # noqa isort:skip
+from .timed_text_track import *  # noqa isort:skip
 from .video import *  # noqa isort:skip
 from .xapi import *  # noqa isort:skip
