+++ src/backend/marsha/core/api/thumbnail.py
@@ -1,0 +1,90 @@
+"""Declare API endpoints for thumbnails with Django RestFramework viewsets."""
+from django.conf import settings
+from django.utils import timezone
+
+from rest_framework import mixins, viewsets
+from rest_framework.decorators import action
+from rest_framework.response import Response
+from rest_framework_simplejwt.models import TokenUser
+
+from .. import defaults, permissions, serializers
+from ..models import Thumbnail
+from ..utils.s3_utils import create_presigned_post
+from ..utils.time_utils import to_timestamp
+from .base import ObjectPkMixin
+
+
+class ThumbnailViewSet(
+    ObjectPkMixin,
+    mixins.CreateModelMixin,
+    mixins.DestroyModelMixin,
+    mixins.RetrieveModelMixin,
+    viewsets.GenericViewSet,
+):
+    """Viewset for the API of the Thumbnail object."""
+
+    permission_classes = [permissions.NotAllowed]
+    serializer_class = serializers.ThumbnailSerializer
+
+    def get_permissions(self):
+        """Instantiate and return the list of permissions that this view requires."""
+        if self.action == "create":
+            permission_classes = [
+                permissions.IsTokenInstructor | permissions.IsTokenAdmin
+            ]
+        else:
+            permission_classes = [
+                permissions.IsTokenResourceRouteObjectRelatedVideo
+                & permissions.IsTokenInstructor
+                | permissions.IsTokenResourceRouteObjectRelatedVideo
+                & permissions.IsTokenAdmin
+            ]
+        return [permission() for permission in permission_classes]
+
+    def get_queryset(self):
+        """Restrict list access to thumbnail related to the video in the JWT token."""
+        user = self.request.user
+        if isinstance(user, TokenUser):
+            return Thumbnail.objects.filter(video__id=user.id)
+        return Thumbnail.objects.none()
+
+    @action(methods=["post"], detail=True, url_path="initiate-upload")
+    # pylint: disable=unused-argument
+    def initiate_upload(self, request, pk=None):
+        """Get an upload policy for a thumbnail.
+
+        Calling the endpoint resets the upload state to `pending` and returns an upload policy to
+        our AWS S3 source bucket.
+
+        Parameters
+        ----------
+        request : Type[django.http.request.HttpRequest]
+            The request on the API endpoint
+        pk: string
+            The primary key of the thumbnail
+
+        Returns
+        -------
+        Type[rest_framework.response.Response]
+            HttpResponse carrying the AWS S3 upload policy as a JSON object.
+
+        """
+        now = timezone.now()
+        stamp = to_timestamp(now)
+
+        thumbnail = self.get_object()
+        key = thumbnail.get_source_s3_key(stamp=stamp)
+
+        presigned_post = create_presigned_post(
+            [
+                ["starts-with", "$Content-Type", "image/"],
+                ["content-length-range", 0, settings.THUMBNAIL_SOURCE_MAX_SIZE],
+            ],
+            {},
+            key,
+        )
+
+        # Reset the upload state of the thumbnail
+        Thumbnail.objects.filter(pk=pk).update(upload_state=defaults.PENDING)
+
+        return Response(presigned_post)
